requirements.txt
openai
python-dotenv
numpy
requests
django
pillow
matplotlib
